Environments/Production/Docker/flask/requirements.txt
flask==3.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
python-dotenv==1.0.0
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
urllib3==1.25.8
paramiko==3.4.0